+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+This project uses [Semantic Versioning 2.0.0](http://semver.org/).
+
+## master
+
+- CHANGED: `Certificate.ExpiresOn` has been replaced by `Certificate.ExpiresAt`. (dnsimple/dnsimple-csharp#14)
+- CHANGED: `Domain.ExpiresOn` has been replaced by `Domain.ExpiresAt`. (dnsimple/dnsimple-csharp#11)
+
 ## Release 0.9.0
 
 Initial non-development release that supports API v2.